Address code review issues: fix train.py redundancy, add tile_overlap validation, fix redundant uint8 cast, add CI permissions
Agent-Logs-Url: https://github.com/fusselc/esrgan-super-resolution/sessions/8fb6be84-f4af-4e16-b322-fd221b4daa4b

+++ src/utils.py
@@ -76,7 +76,7 @@ def save_image(tensor: torch.Tensor, path: str) -> None:
     if tensor.ndim == 4:
         tensor = tensor[0]
     arr = tensor_to_numpy(tensor)
-    img = Image.fromarray(arr.astype(np.uint8))
+    img = Image.fromarray(arr)
     img.save(path)
